requirements.txt
typer~=0.9.0
python-dotenv~=1.0.0
openai==1.12.0
loguru~=0.7.2
neo4j==5.17.0
tiktoken==0.6.0
torch~=2.2.1
huggingface-hub~=0.21.1
numpy~=1.26.4
pillow~=10.2.0
torchvision~=0.17.1